+++ Changelog.md
@@ -17,6 +17,7 @@ Compiler Features:
  * SMTChecker: Show unsupported warning for inline assembly blocks.
  * SMTChecker: Support mod.
  * SMTChecker: Support ``contract`` type.
+ * SMTChecker: Support ``this`` as address.
  * Optimizer: Add rule for shifts by constants larger than 255 for Constantinople.
  * Optimizer: Add rule to simplify certain ANDs and SHL combinations
  * Yul: Adds break and continue keywords to for-loop syntax.